feat: Version 1.x of totally-untappd
Totally-Untappd has now moved to a new architecture!

The Total Wine scraper has moved to a separate file which handles
a lot of the return codes from the requests module and attempts
automatic retrying.
It also uses the random users agent module to ensure we aren't detected
too quickly.

The Untappd scraper also handles return codes from the requests module
as well, and attempts automatic retrying in the case of a 429 retry code.
Unfortunately, this makes it very slow since the typically retry-after
is 65 seconds long after approximately ~90 requests.

The db.py shows how sqlite3 database is organized in terms of tables and
views. The main.py file uses this schema to store the data generated
by Untappd and Total Wine.

Signed-off-by: Ryan Sherlock <ryan.m.sherlock@gmail.com>